Update README.md
This is an automatic trader using Alpaca's free paper trading API. It uses Bollinger Bands which takes the average of a certain symbols stock price of the past 20 days from the current day, and forms upper and lower limits that are both 2 standard deviations above and below the calculated average. When the stock price rises above the upper limit, the program will see this as an indicator to sell whatever shares the user owns of the corresponding symbol. When the stock price falls below the lower limit, the program will see this as an indicator to buy whatever shares the user can buy of the corresponding symbol.